devices: doc for the balloon device
