+++ src/devices/src/virtio/balloon/device.rs
@@ -15,18 +15,19 @@ use utils::eventfd::EventFd;
 use virtio_gen::virtio_blk::VIRTIO_F_VERSION_1;
 use vm_memory::{Address, ByteValued, Bytes, GuestAddress, GuestMemoryMmap};
 
-use super::super::{ActivateResult, DeviceState, Queue, VirtioDevice, TYPE_BALLOON};
-use super::utils::{compact_page_frame_numbers, remove_range};
-use super::{
-    BALLOON_DEV_ID, DEFLATE_INDEX, INFLATE_INDEX, MAX_PAGES_IN_DESC, MAX_PAGE_COMPACT_BUFFER,
-    MIB_TO_4K_PAGES, NUM_QUEUES, QUEUE_SIZES, STATS_INDEX, VIRTIO_BALLOON_F_DEFLATE_ON_OOM,
-    VIRTIO_BALLOON_F_STATS_VQ, VIRTIO_BALLOON_PFN_SHIFT, VIRTIO_BALLOON_S_AVAIL,
-    VIRTIO_BALLOON_S_CACHES, VIRTIO_BALLOON_S_HTLB_PGALLOC, VIRTIO_BALLOON_S_HTLB_PGFAIL,
-    VIRTIO_BALLOON_S_MAJFLT, VIRTIO_BALLOON_S_MEMFREE, VIRTIO_BALLOON_S_MEMTOT,
-    VIRTIO_BALLOON_S_MINFLT, VIRTIO_BALLOON_S_SWAP_IN, VIRTIO_BALLOON_S_SWAP_OUT,
+use crate::virtio::balloon::utils::{compact_page_frame_numbers, remove_range};
+use crate::virtio::balloon::{
+    Error as BalloonError, BALLOON_DEV_ID, DEFLATE_INDEX, INFLATE_INDEX, MAX_PAGES_IN_DESC,
+    MAX_PAGE_COMPACT_BUFFER, MIB_TO_4K_PAGES, NUM_QUEUES, QUEUE_SIZES, STATS_INDEX,
+    VIRTIO_BALLOON_F_DEFLATE_ON_OOM, VIRTIO_BALLOON_F_STATS_VQ, VIRTIO_BALLOON_PFN_SHIFT,
+    VIRTIO_BALLOON_S_AVAIL, VIRTIO_BALLOON_S_CACHES, VIRTIO_BALLOON_S_HTLB_PGALLOC,
+    VIRTIO_BALLOON_S_HTLB_PGFAIL, VIRTIO_BALLOON_S_MAJFLT, VIRTIO_BALLOON_S_MEMFREE,
+    VIRTIO_BALLOON_S_MEMTOT, VIRTIO_BALLOON_S_MINFLT, VIRTIO_BALLOON_S_SWAP_IN,
+    VIRTIO_BALLOON_S_SWAP_OUT,
 };
-use crate::virtio::balloon::Error as BalloonError;
-use crate::virtio::{IrqTrigger, IrqType};
+use crate::virtio::device::{DeviceState, IrqTrigger, IrqType, VirtioDevice};
+use crate::virtio::queue::Queue;
+use crate::virtio::{ActivateResult, TYPE_BALLOON};
 
 const SIZE_OF_U32: usize = std::mem::size_of::<u32>();
 const SIZE_OF_STAT: usize = std::mem::size_of::<BalloonStat>();
@@ -63,40 +64,63 @@ struct BalloonStat {
 // SAFETY: Safe because BalloonStat only contains plain data.
 unsafe impl ByteValued for BalloonStat {}
 
-// BalloonStats holds statistics returned from the stats_queue.
+/// Holds configuration details for the device.
 #[derive(Clone, Default, Debug, PartialEq, Eq, Serialize)]
 pub struct BalloonConfig {
+    /// Target size.
     pub amount_mib: u32,
+    /// Whether or not to ask for pages back.
     pub deflate_on_oom: bool,
+    /// Interval of time in seconds at which the balloon statistics are updated.
     pub stats_polling_interval_s: u16,
 }
 
-// BalloonStats holds statistics returned from the stats_queue.
+/// BalloonStats holds statistics returned from the stats_queue.
 #[derive(Clone, Default, Debug, PartialEq, Eq, Serialize)]
 #[serde(deny_unknown_fields)]
 pub struct BalloonStats {
+    /// The target size of the balloon, in 4K pages.
     pub target_pages: u32,
+    /// The number of 4K pages the device is currently holding.
     pub actual_pages: u32,
+    /// The target size of the balloon, in MiB.
     pub target_mib: u32,
+    /// The number of MiB the device is currently holding.
     pub actual_mib: u32,
+    /// Amount of memory swapped in.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub swap_in: Option<u64>,
+    /// Amount of memory swapped out.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub swap_out: Option<u64>,
+    /// Number of major faults.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub major_faults: Option<u64>,
+    /// Number of minor faults.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub minor_faults: Option<u64>,
+    /// The amount of memory not being used for any
+    ///  purpose (in bytes).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub free_memory: Option<u64>,
+    /// Total amount of memory available (in bytes).
     #[serde(skip_serializing_if = "Option::is_none")]
     pub total_memory: Option<u64>,
+    /// An estimate of how much memory is available (in
+    ///  bytes) for starting new applications, without pushing the system to swap.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub available_memory: Option<u64>,
+    /// The amount of memory, in bytes, that can be
+    ///   quickly reclaimed without additional I/O. Typically these pages are used for
+    ///   caching files from disk.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub disk_caches: Option<u64>,
+    /// The number of successful hugetlb page
+    /// allocations in the guest.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub hugetlb_allocations: Option<u64>,
+    /// The number of failed hugetlb page allocations
+    ///  in the guest.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub hugetlb_failures: Option<u64>,
 }
@@ -124,7 +148,7 @@ impl BalloonStats {
     }
 }
 
-// Virtio balloon device.
+/// Virtio balloon device.
 pub struct Balloon {
     // Virtio fields.
     pub(crate) avail_features: u64,
@@ -151,6 +175,7 @@ pub struct Balloon {
 }
 
 impl Balloon {
+    /// Instantiate a new balloon device.
     pub fn new(
         amount_mib: u32,
         deflate_on_oom: bool,
@@ -395,6 +420,7 @@ impl Balloon {
         let _ = self.process_deflate_queue();
     }
 
+    /// Provides the ID of this balloon device.
     pub fn id(&self) -> &str {
         BALLOON_DEV_ID
     }
@@ -416,6 +442,7 @@ impl Balloon {
         }
     }
 
+    /// Update the target size of the balloon.
     pub fn update_size(&mut self, amount_mib: u32) -> Result<(), BalloonError> {
         if self.is_activated() {
             self.config_space.num_pages = mib_to_pages(amount_mib)?;
@@ -427,6 +454,7 @@ impl Balloon {
         }
     }
 
+    /// Update the the statistics polling interval.
     pub fn update_stats_polling_interval(&mut self, interval_s: u16) -> Result<(), BalloonError> {
         if self.stats_polling_interval_s == interval_s {
             return Ok(());
@@ -443,7 +471,7 @@ impl Balloon {
         Ok(())
     }
 
-    pub fn update_timer_state(&mut self) {
+    fn update_timer_state(&mut self) {
         let timer_state = TimerState::Periodic {
             current: Duration::from_secs(u64::from(self.stats_polling_interval_s)),
             interval: Duration::from_secs(u64::from(self.stats_polling_interval_s)),
@@ -452,22 +480,25 @@ impl Balloon {
             .set_state(timer_state, SetTimeFlags::Default);
     }
 
+    /// Obtain the number of 4K pages the device is currently holding.
     pub fn num_pages(&self) -> u32 {
         self.config_space.num_pages
     }
 
+    /// Obtain the size of 4K pages the device is currently holding in MIB.
     pub fn size_mb(&self) -> u32 {
         pages_to_mib(self.config_space.num_pages)
     }
 
-    pub fn deflate_on_oom(&self) -> bool {
+    fn deflate_on_oom(&self) -> bool {
         self.avail_features & (1u64 << VIRTIO_BALLOON_F_DEFLATE_ON_OOM) != 0
     }
 
-    pub fn stats_polling_interval_s(&self) -> u16 {
+    fn stats_polling_interval_s(&self) -> u16 {
         self.stats_polling_interval_s
     }
 
+    /// Retrieve latest stats for the balloon device.
     pub fn latest_stats(&mut self) -> Option<&BalloonStats> {
         if self.stats_enabled() {
             self.latest_stats.target_pages = self.config_space.num_pages;
@@ -480,6 +511,7 @@ impl Balloon {
         }
     }
 
+    /// Return the config of the balloon device.
     pub fn config(&self) -> BalloonConfig {
         BalloonConfig {
             amount_mib: self.size_mb(),
